Corrected LCOSS DIS switch description (#441)
* Corrected LCOSS DIS switch description

The LCOSS DIS switch was part of the APS-107 integration and likely switched if the gunsight glass would show info from the gunsight system (originally ASG-26A in the F-4E) or the APS-107 RWR. 

The ARS-107 is likely a typo and ment to say APS-107. This RHAW system was able to project RWR data from four front facing antennas on the DVST screen or the gunsight glass. The DVST description was already fixed.

This info comes mainly from this discussion and additional recherches on the discord:
https://groups.google.com/g/rec.aviation.military/c/x0o9KYOVHmI

For detail, this RHAW system had the ability to display information from four front facing antennas on DVST scope/LCOSS glass. This was likely a feature to help guide AGM-78s into the target, a central feature of the APS-107 system that was largely abandoned at some point.

* Fixed trailing spaces. Hopefully!